Properly respect allow judge when creating judge tasks.
Note that this is only done completely for language for now and still
needs to implemented for problem.

As discussed on https://docs.google.com/document/d/1dMjPOcR2TqS7Z4jNbb5W5vVdHu43BSf79gRngEkfk5w/edit?disco=AAAAGdSchic
we do create judgetasks eagerly. The consequence of this is that when
judging is disabled for a language, we stop creating judgetasks. When
later judging is enabled again (e.g. after fixing the compile script),
we need to create the judgetasks at that moment.

This comes with a little confusing ergonomics. In the master branch, if
you disallow judging on a language, it won't start any judging of
submissions with that language. In this branch, it will still judge
(likely incorrectly) all pending (but not future!) submissions with that
language. There's a TODO in the JudgehostController to potentially
disable other judgings to bring back the old behavior (which is IMHO
more obvious and also less wasteful).